ops/aiops/requirements.txt
requests>=2.28
PyYAML>=6.0
GitPython>=3.1
pynacl>=1.5
